refactor: 알림설정 Request DTO 패키지 정리
